test_project/requirements.txt
kindwise-api-client[async]==0.8.1
python-dotenv
